Refactor environment imports to use the new manipulation module and remove obsolete Openloong example scripts
