+++ examples/imitation/run_franka_single_arm.py
@@ -17,7 +17,7 @@ from datetime import datetime
 from orca_gym.environment.orca_gym_env import RewardType
 from orca_gym.robomimic.dataset_util import DatasetWriter, DatasetReader
 from orca_gym.sensor.rgbd_camera import Monitor, CameraWrapper
-from envs.franka.franka_env import FrankaEnv, RunMode, ControlDevice
+from envs.manipulation.franka_env import FrankaEnv, RunMode, ControlDevice
 from examples.imitation.train_policy import train_policy
 from examples.imitation.test_policy import create_env, rollout
 from orca_gym.utils.dir_utils import create_tmp_dir
