--- examples/openloong/Openloong_arm_xbox_osc_ctrl.py
@@ -1,69 +1,0 @@
-import os
-import sys
-import time
-
-current_file_path = os.path.abspath('./../../')
-
-if current_file_path not in sys.path:
-    print("add path: ", current_file_path)
-    sys.path.append(current_file_path)
-
-
-import gymnasium as gym
-from gymnasium.envs.registration import register
-from datetime import datetime
-
-# 
-TIME_STEP = 0.01
-
-def register_env(orcagym_addr, env_name, env_index, control_freq=20) -> str:
-    orcagym_addr_str = orcagym_addr.replace(":", "-")
-    env_id = env_name + "-OrcaGym-" + orcagym_addr_str + f"-{env_index:03d}"
-    gym.register(
-        id=env_id,
-        entry_point="envs.openloong.Openloong_arm_xbox_env:OpenloongArmEnv",
-        kwargs={'frame_skip': 1,   
-                'reward_type': "dense",
-                'orcagym_addr': orcagym_addr, 
-                'agent_names': ['AzureLoong'], 
-                'time_step': TIME_STEP,
-                'control_freq': control_freq},
-        max_episode_steps=sys.maxsize,
-        reward_threshold=0.0,
-    )
-    return env_id
-
-def continue_training(env):
-    observation, info = env.reset(seed=42)
-    while True:
-        start_time = datetime.now()
-
-        action = env.action_space.sample()
-        observation, reward, terminated, truncated, info = env.step(action)
-
-        elapsed_time = datetime.now() - start_time
-        if elapsed_time.total_seconds() < TIME_STEP:
-            time.sleep(TIME_STEP - elapsed_time.total_seconds())
-
-    
-
-if __name__ == "__main__":
-    """
-    OSC运动算法控制青龙机器人机械臂的示例
-    """
-    try:
-        orcagym_addr = "localhost:50051"
-        print("simulation running... , orcagym_addr: ", orcagym_addr)
-
-        env_name = "OpenloongArmXbox-v0"
-        env_index = 0
-        env_id = register_env(orcagym_addr, env_name, env_index, 20)
-        print("Registering environment with id: ", env_id)
-
-        env = gym.make(env_id)        
-        print("Starting simulation...")
-
-        continue_training(env)
-    except KeyboardInterrupt:
-        print("Simulation stopped")        
-        env.close()
