+++ orca_gym/scripts/franka_manipulation.py
@@ -12,7 +12,7 @@ from datetime import datetime
 from orca_gym.environment.orca_gym_env import RewardType
 from orca_gym.robomimic.dataset_util import DatasetWriter, DatasetReader
 from orca_gym.sensor.rgbd_camera import Monitor, CameraWrapper
-from envs.franka.franka_env import FrankaEnv, RunMode, ControlDevice
+from envs.manipulation.franka_env import FrankaEnv, RunMode, ControlDevice
 from examples.imitation.train_policy import train_policy
 from examples.imitation.test_policy import create_env, rollout
 from orca_gym.utils.dir_utils import create_tmp_dir
@@ -23,7 +23,7 @@ import numpy as np
 
 
 ENV_ENTRY_POINT = {
-    "Franka": "envs.franka.franka_env:FrankaEnv"
+    "Franka": "envs.manipulation.franka_env:FrankaEnv"
 }
 
 TIME_STEP = 0.005
